start with refactoring
